Reorganize genetic value headers to better match Python names and to separate based on needed includes. Closes #308.
