+++ fwdpy11/headers/fwdpy11/genetic_value_to_fitness/MultivariateGSSmo.hpp
@@ -24,7 +24,7 @@
 #include <cmath>
 #include <vector>
 #include <stdexcept>
-#include "GeneticValueToFitnessMap.hpp"
+#include "GeneticValueIsTrait.hpp"
 
 namespace fwdpy11
 {
