+++ fwdpy11/src/genetic_values/DiploidGeneticValue.cc
@@ -1,5 +1,6 @@
 #include <fwdpy11/genetic_values/DiploidGeneticValue.hpp>
 #include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 
 namespace py = pybind11;
